Moved HttpClientConfigurator and RequestBuilder to Mailgun\HttpClient namespace
